requirements.txt
pillow
qrcode==6.0
reportlab==3.4.0
